+++ src/skill_ids.h
@@ -1,151 +1,151 @@
 #pragma once
 
 //vale guardian
-const uint16_t MECHANIC_VG_TELEPORT_GREEN = 31860;//Unstable Magic Spike
-const uint16_t MECHANIC_VG_TELEPORT_RAINBOW = 31392;//Unstable Magic Spike
+const uint32_t MECHANIC_VG_TELEPORT_GREEN = 31860;//Unstable Magic Spike
+const uint32_t MECHANIC_VG_TELEPORT_RAINBOW = 31392;//Unstable Magic Spike
 
-const uint16_t MECHANIC_VG_GREEN_A = 31340;//I'm not sure why this mechanic has 4 ids, but it appears to for some reason
-const uint16_t MECHANIC_VG_GREEN_B = 31391;//all these ids are for when 4 people are in the green circle
-const uint16_t MECHANIC_VG_GREEN_C = 31529;//it appears to be a separate id for the 90% hp blast when <4 people are in the green
-const uint16_t MECHANIC_VG_GREEN_D = 31750;//all 4 ids are called "Distributed Magic"
+const uint32_t MECHANIC_VG_GREEN_A = 31340;//I'm not sure why this mechanic has 4 ids, but it appears to for some reason
+const uint32_t MECHANIC_VG_GREEN_B = 31391;//all these ids are for when 4 people are in the green circle
+const uint32_t MECHANIC_VG_GREEN_C = 31529;//it appears to be a separate id for the 90% hp blast when <4 people are in the green
+const uint32_t MECHANIC_VG_GREEN_D = 31750;//all 4 ids are called "Distributed Magic"
 
 //gorseval
-const uint16_t MECHANIC_GORS_SLAM = 31875;//Spectral Impact
-const uint16_t MECHANIC_GORS_EGG = 31623;//Ghastly Prison
-const uint16_t MECHANIC_GORS_ORB = 31498;//Spectral Darkness
+const uint32_t MECHANIC_GORS_SLAM = 31875;//Spectral Impact
+const uint32_t MECHANIC_GORS_EGG = 31623;//Ghastly Prison
+const uint32_t MECHANIC_GORS_ORB = 31498;//Spectral Darkness
 
 //sabetha
-const uint16_t MECHANIC_SAB_SAPPER_BOMB = 31473;//Sapper Bomb
-const uint16_t MECHANIC_SAB_TIME_BOMB = 31485;//Time Bomb
-const uint16_t MECHANIC_SAB_CANNON = 31643;//Cannon Barrage
-const uint16_t MECHANIC_SAB_FLAMEWALL = 31332;//Firestorm
+const uint32_t MECHANIC_SAB_SAPPER_BOMB = 31473;//Sapper Bomb
+const uint32_t MECHANIC_SAB_TIME_BOMB = 31485;//Time Bomb
+const uint32_t MECHANIC_SAB_CANNON = 31643;//Cannon Barrage
+const uint32_t MECHANIC_SAB_FLAMEWALL = 31332;//Firestorm
 
 //slothasor
-const uint16_t MECHANIC_SLOTH_TANTRUM = 34479;//Tantrum
-const uint16_t MECHANIC_SLOTH_BOMB = 34387;//Volatile Poison
-const uint16_t MECHANIC_SLOTH_BOMB_AOE = 34481;//Volatile Poison
-const uint16_t MECHANIC_SLOTH_FLAME_BREATH = 34516;//nameless
-const uint16_t MECHANIC_SLOTH_SHAKE = 34482;//nameless
-const uint16_t MECHANIC_SLOTH_FIXATE = 34508;//Fixated
+const uint32_t MECHANIC_SLOTH_TANTRUM = 34479;//Tantrum
+const uint32_t MECHANIC_SLOTH_BOMB = 34387;//Volatile Poison
+const uint32_t MECHANIC_SLOTH_BOMB_AOE = 34481;//Volatile Poison
+const uint32_t MECHANIC_SLOTH_FLAME_BREATH = 34516;//nameless
+const uint32_t MECHANIC_SLOTH_SHAKE = 34482;//nameless
+const uint32_t MECHANIC_SLOTH_FIXATE = 34508;//Fixated
 
 //matthias
-const uint16_t MECHANIC_MATT_HADOUKEN_HUMAN = 34380;//Oppressive Gaze
-const uint16_t MECHANIC_MATT_HADOUKEN_ABOM = 34371;//Oppressive Gaze
-const uint16_t MECHANIC_MATT_SHARD_HUMAN = 34404;//Shards of Rage
-const uint16_t MECHANIC_MATT_SHARD_ABOM = 34411;//Shards of Rage
-const uint16_t MECHANIC_MATT_BOMB = 34450;//Unstable Blood Magic
-const uint16_t MECHANIC_MATT_CORRUPTION = 34416;//Corruption
-const uint16_t MECHANIC_MATT_SACRIFICE = 34442;//nameless
+const uint32_t MECHANIC_MATT_HADOUKEN_HUMAN = 34380;//Oppressive Gaze
+const uint32_t MECHANIC_MATT_HADOUKEN_ABOM = 34371;//Oppressive Gaze
+const uint32_t MECHANIC_MATT_SHARD_HUMAN = 34404;//Shards of Rage
+const uint32_t MECHANIC_MATT_SHARD_ABOM = 34411;//Shards of Rage
+const uint32_t MECHANIC_MATT_BOMB = 34450;//Unstable Blood Magic
+const uint32_t MECHANIC_MATT_CORRUPTION = 34416;//Corruption
+const uint32_t MECHANIC_MATT_SACRIFICE = 34442;//nameless
 
 //kc
-const uint16_t MECHANIC_KC_FIXATE_EAST = 34912;//Fixated
-const uint16_t MECHANIC_KC_FIXATE_WEST = 34925;//Fixated
+const uint32_t MECHANIC_KC_FIXATE_EAST = 34912;//Fixated
+const uint32_t MECHANIC_KC_FIXATE_WEST = 34925;//Fixated
 
 //xera
-const uint16_t MECHANIC_XERA_HALF = 34883;
-const uint16_t MECHANIC_XERA_ORB = 35128;//Temporal Shred
-const uint16_t MECHANIC_XERA_ORB_AOE = 34913;//Temporal Shred
-const uint16_t MECHANIC_XERA_MAGIC = 35000;//Intervention
+const uint32_t MECHANIC_XERA_HALF = 34883;
+const uint32_t MECHANIC_XERA_ORB = 35128;//Temporal Shred
+const uint32_t MECHANIC_XERA_ORB_AOE = 34913;//Temporal Shred
+const uint32_t MECHANIC_XERA_MAGIC = 35000;//Intervention
 
 //carin
-const uint16_t MECHANIC_CARIN_TELEPORT = 38113;//Displacement
-const uint16_t MECHANIC_CARIN_SHARD = 38313;//Meteor Swarm
-const uint16_t MECHANIC_CARIN_SWEEP = 37631;//Orbital Sweep
-const uint16_t MECHANIC_CARIN_GREEN_A = 37611;//Spatial Manipulation, 1 id per pattern of greens
-const uint16_t MECHANIC_CARIN_GREEN_B = 37629;//Spatial Manipulation
-const uint16_t MECHANIC_CARIN_GREEN_C = 37642;//Spatial Manipulation
-const uint16_t MECHANIC_CARIN_GREEN_D = 37673;//Spatial Manipulation
-const uint16_t MECHANIC_CARIN_GREEN_E = 38074;//Spatial Manipulation
-const uint16_t MECHANIC_CARIN_GREEN_F = 38302;//Spatial Manipulation
+const uint32_t MECHANIC_CARIN_TELEPORT = 38113;//Displacement
+const uint32_t MECHANIC_CARIN_SHARD = 38313;//Meteor Swarm
+const uint32_t MECHANIC_CARIN_SWEEP = 37631;//Orbital Sweep
+const uint32_t MECHANIC_CARIN_GREEN_A = 37611;//Spatial Manipulation, 1 id per pattern of greens
+const uint32_t MECHANIC_CARIN_GREEN_B = 37629;//Spatial Manipulation
+const uint32_t MECHANIC_CARIN_GREEN_C = 37642;//Spatial Manipulation
+const uint32_t MECHANIC_CARIN_GREEN_D = 37673;//Spatial Manipulation
+const uint32_t MECHANIC_CARIN_GREEN_E = 38074;//Spatial Manipulation
+const uint32_t MECHANIC_CARIN_GREEN_F = 38302;//Spatial Manipulation
 
 //sam
-const uint16_t MECHANIC_SAM_SHOCKWAVE = 37996;
-const uint16_t MECHANIC_SAM_SLAP_HORIZONTAL = 38168;
-const uint16_t MECHANIC_SAM_SLAP_VERTICAL = 38305;//Bludgeon
-const uint16_t MECHANIC_SAM_FIXATE_SAM = 37868;//Fixate: Samarog
-const uint16_t MECHANIC_SAM_GREEN_BIG = 37966;//unnamed
-const uint16_t MECHANIC_SAM_GREEN_SMALL = 38247;//unnamed
+const uint32_t MECHANIC_SAM_SHOCKWAVE = 37996;
+const uint32_t MECHANIC_SAM_SLAP_HORIZONTAL = 38168;
+const uint32_t MECHANIC_SAM_SLAP_VERTICAL = 38305;//Bludgeon
+const uint32_t MECHANIC_SAM_FIXATE_SAM = 37868;//Fixate: Samarog
+const uint32_t MECHANIC_SAM_GREEN_BIG = 37966;//unnamed
+const uint32_t MECHANIC_SAM_GREEN_SMALL = 38247;//unnamed
 
 //deimos
-const uint16_t MECHANIC_DEIMOS_OIL = 37716;
-const uint16_t MECHANIC_DEIMOS_SMASH = 38208; //smash from the actual boss
-const uint16_t MECHANIC_DEIMOS_SMASH_INITIAL = 37929; //smash while killing the initial chains
-const uint16_t MECHANIC_DEIMOS_SMASH_END_A = 37980;//Demonic Shock Wave,smash during last 10%
-const uint16_t MECHANIC_DEIMOS_SMASH_END_B = 37982;//different skill id for the 2 sides of the platform
-const uint16_t MECHANIC_DEIMOS_TEAR = 37733;//Tear Instability, buff from closing a tear
+const uint32_t MECHANIC_DEIMOS_OIL = 37716;
+const uint32_t MECHANIC_DEIMOS_SMASH = 38208; //smash from the actual boss
+const uint32_t MECHANIC_DEIMOS_SMASH_INITIAL = 37929; //smash while killing the initial chains
+const uint32_t MECHANIC_DEIMOS_SMASH_END_A = 37980;//Demonic Shock Wave,smash during last 10%
+const uint32_t MECHANIC_DEIMOS_SMASH_END_B = 37982;//different skill id for the 2 sides of the platform
+const uint32_t MECHANIC_DEIMOS_TEAR = 37733;//Tear Instability, buff from closing a tear
 
 //Soulless Horror
-const uint16_t MECHANIC_HORROR_DONUT_INNER = 47327;//Vortex Slash
-const uint16_t MECHANIC_HORROR_DONUT_OUTER = 48432;//Vortex Slash ingame, achieve claims it's called "hurricane slash", but it's not
-const uint16_t MECHANIC_HORROR_GOLEM_AOE = 47430;//Soul Rift
-const uint16_t MECHANIC_HORROR_PIE_4_A = 48363;//Quad Slash
-const uint16_t MECHANIC_HORROR_PIE_4_B = 47915;//Quad Slash
-const uint16_t MECHANIC_HORROR_SCYTHE = 47363;//Spinning Slash
-const uint16_t MECHANIC_HORROR_FIXATE = 47434;//Fixated
-const uint16_t MECHANIC_HORROR_DEBUFF = 47414;//Necrosis
+const uint32_t MECHANIC_HORROR_DONUT_INNER = 47327;//Vortex Slash
+const uint32_t MECHANIC_HORROR_DONUT_OUTER = 48432;//Vortex Slash ingame, achieve claims it's called "hurricane slash", but it's not
+const uint32_t MECHANIC_HORROR_GOLEM_AOE = 47430;//Soul Rift
+const uint32_t MECHANIC_HORROR_PIE_4_A = 48363;//Quad Slash
+const uint32_t MECHANIC_HORROR_PIE_4_B = 47915;//Quad Slash
+const uint32_t MECHANIC_HORROR_SCYTHE = 47363;//Spinning Slash
+const uint32_t MECHANIC_HORROR_FIXATE = 47434;//Fixated
+const uint32_t MECHANIC_HORROR_DEBUFF = 47414;//Necrosis
 
 //Eater of Souls
-const uint16_t MECHANIC_EATER_PUKE = 47303;//Hungering Miasma
-const uint16_t MECHANIC_EATER_WEB = 47695;//Spinner's Web
+const uint32_t MECHANIC_EATER_PUKE = 47303;//Hungering Miasma
+const uint32_t MECHANIC_EATER_WEB = 47695;//Spinner's Web
 
 //Dhuum
-const uint16_t MECHANIC_DHUUM_GOLEM = 48172; //Hateful Ephemera (flesh golem)
-const uint16_t MECHANIC_DHUUM_AFFLICTION = 47646;//Arcing Affliction
-const uint16_t MECHANIC_DHUUM_SHACKLE = 47335;//Soul Shackle
-const uint16_t MECHANIC_DHUUM_CRACK = 48752;//Cull
-const uint16_t MECHANIC_DHUUM_MARK = 48760;//Putrid Bomb
-const uint16_t MECHANIC_DHUUM_SUCK_AOE = 48398;//Cataclysmic Cycle
-const uint16_t MECHANIC_DHUUM_TELEPORT_AOE = 48176;//Death Mark
-const uint16_t MECHANIC_DHUUM_GREEN_TIMER = 48281;//Mortal Coil
-const uint16_t MECHANIC_DHUUM_SNATCH = 47076;//nameless
-const uint16_t MECHANIC_DHUUM_BUTTON_CHANNEL = 47002;//nameless
-const uint16_t MECHANIC_DHUUM_CONE = 47561;//Slash
+const uint32_t MECHANIC_DHUUM_GOLEM = 48172; //Hateful Ephemera (flesh golem)
+const uint32_t MECHANIC_DHUUM_AFFLICTION = 47646;//Arcing Affliction
+const uint32_t MECHANIC_DHUUM_SHACKLE = 47335;//Soul Shackle
+const uint32_t MECHANIC_DHUUM_CRACK = 48752;//Cull
+const uint32_t MECHANIC_DHUUM_MARK = 48760;//Putrid Bomb
+const uint32_t MECHANIC_DHUUM_SUCK_AOE = 48398;//Cataclysmic Cycle
+const uint32_t MECHANIC_DHUUM_TELEPORT_AOE = 48176;//Death Mark
+const uint32_t MECHANIC_DHUUM_GREEN_TIMER = 48281;//Mortal Coil
+const uint32_t MECHANIC_DHUUM_SNATCH = 47076;//nameless
+const uint32_t MECHANIC_DHUUM_BUTTON_CHANNEL = 47002;//nameless
+const uint32_t MECHANIC_DHUUM_CONE = 47561;//Slash
 
-const uint16_t MECHANIC_NIGHTMARE_VOMIT = 37030;//Toxic Sickness
+const uint32_t MECHANIC_NIGHTMARE_VOMIT = 37030;//Toxic Sickness
 
 //MAMA
-const uint16_t MECHANIC_MAMA_WIRL = 37408;//Blastwave
-const uint16_t MECHANIC_MAMA_WIRL_NORMAL = 37103;//unnamed, wirl on normal mode
-const uint16_t MECHANIC_MAMA_KNOCK = 37391;//Tantrum
-const uint16_t MECHANIC_MAMA_LEAP = 37577;//Leap
-const uint16_t MECHANIC_MAMA_ACID = 37437;//Shoot
-const uint16_t MECHANIC_MAMA_KNIGHT_SMASH = 37185;//Explosive Impact
+const uint32_t MECHANIC_MAMA_WIRL = 37408;//Blastwave
+const uint32_t MECHANIC_MAMA_WIRL_NORMAL = 37103;//unnamed, wirl on normal mode
+const uint32_t MECHANIC_MAMA_KNOCK = 37391;//Tantrum
+const uint32_t MECHANIC_MAMA_LEAP = 37577;//Leap
+const uint32_t MECHANIC_MAMA_ACID = 37437;//Shoot
+const uint32_t MECHANIC_MAMA_KNIGHT_SMASH = 37185;//Explosive Impact
 
 //Siax
-const uint16_t MECHANIC_SIAX_ACID = 37477;//Vile Spit
+const uint32_t MECHANIC_SIAX_ACID = 37477;//Vile Spit
 
 //Ensolyss
-const uint16_t MECHANIC_ENSOLYSS_LUNGE = 37154;//Lunge
-const uint16_t MECHANIC_ENSOLYSS_SMASH_INITIAL = 37278;//The knockback part
-const uint16_t MECHANIC_ENSOLYSS_SMASH = 36962;//The torment part (up to 5 hits, 1 per player)
+const uint32_t MECHANIC_ENSOLYSS_LUNGE = 37154;//Lunge
+const uint32_t MECHANIC_ENSOLYSS_SMASH_INITIAL = 37278;//The knockback part
+const uint32_t MECHANIC_ENSOLYSS_SMASH = 36962;//The torment part (up to 5 hits, 1 per player)
 
 //Arkk
-const uint16_t MECHANIC_ARKK_PIE_A = 39685;//Horizon Strike (the id appears to be different for each phase of Arkk)
-const uint16_t MECHANIC_ARKK_PIE_B = 39001;//Horizon Strike
-const uint16_t MECHANIC_ARKK_PIE_C = 39297;//Horizon Strike (normal mode)
-const uint16_t MECHANIC_ARKK_FEAR = 39442;//Blinding Radiance
-const uint16_t MECHANIC_ARKK_OVERHEAD_SMASH = 38844;//Overhead Smash
-const uint16_t MECHANIC_ARKK_BOMB = 38880;//Corporeal Reassignment
-const uint16_t MECHANIC_ARKK_GOOP = 39711;//unnamed,the projectile that spawns the goop
+const uint32_t MECHANIC_ARKK_PIE_A = 39685;//Horizon Strike (the id appears to be different for each phase of Arkk)
+const uint32_t MECHANIC_ARKK_PIE_B = 39001;//Horizon Strike
+const uint32_t MECHANIC_ARKK_PIE_C = 39297;//Horizon Strike (normal mode)
+const uint32_t MECHANIC_ARKK_FEAR = 39442;//Blinding Radiance
+const uint32_t MECHANIC_ARKK_OVERHEAD_SMASH = 38844;//Overhead Smash
+const uint32_t MECHANIC_ARKK_BOMB = 38880;//Corporeal Reassignment
+const uint32_t MECHANIC_ARKK_GOOP = 39711;//unnamed,the projectile that spawns the goop
 
 //conjures
-const uint16_t CONJURE_ICE_BOW = 5567;//Conjure Frost Bow
-const uint16_t CONJURE_ICE_BOW_BUFF = 15790;//Conjure Frost Attributes
+const uint32_t CONJURE_ICE_BOW = 5567;//Conjure Frost Bow
+const uint32_t CONJURE_ICE_BOW_BUFF = 15790;//Conjure Frost Attributes
 
-const uint16_t CONJURE_LIGHTNING_HAMMER = 5624;//Conjure Lightning Hammer
-const uint16_t CONJURE_LIGHTNING_HAMMER_BUFF = 15791;//Conjure Lightning Attributes
+const uint32_t CONJURE_LIGHTNING_HAMMER = 5624;//Conjure Lightning Hammer
+const uint32_t CONJURE_LIGHTNING_HAMMER_BUFF = 15791;//Conjure Lightning Attributes
 
-const uint16_t CONJURE_FLAME_AXE = 5540;//Conjure Flame Axe
-const uint16_t CONJURE_FLAME_AXE_BUFF = 15789;//Conjure Flame Attributes
+const uint32_t CONJURE_FLAME_AXE = 5540;//Conjure Flame Axe
+const uint32_t CONJURE_FLAME_AXE_BUFF = 15789;//Conjure Flame Attributes
 
-const uint16_t CONJURE_EARTH_SHIELD = 5546;//Conjure Earth Shield
-const uint16_t CONJURE_EARTH_SHIELD_BUFF = 15788;//Conjure Earth Attributes
+const uint32_t CONJURE_EARTH_SHIELD = 5546;//Conjure Earth Shield
+const uint32_t CONJURE_EARTH_SHIELD_BUFF = 15788;//Conjure Earth Attributes
 
-const uint16_t CONJURE_FIRE_GS = 5516;//Conjure Fiery Greatsword
-const uint16_t CONJURE_FIRE_GS_BUFF = 15792;//Conjure Fire Attributes
+const uint32_t CONJURE_FIRE_GS = 5516;//Conjure Fiery Greatsword
+const uint32_t CONJURE_FIRE_GS_BUFF = 15792;//Conjure Fire Attributes
 
 //Buffs
-const uint16_t BUFF_STABILITY = 1122;//Stability
-const uint16_t BUFF_VAPOR_FORM = 5620;//Vapor Form
-const uint16_t BUFF_ILLUSION_OF_LIFE = 10346;//Illusion of Life
-const uint16_t BUFF_CRYSTALLINE_HEART = 37657;//Crystalline Heart, for envoy armor 2 collection
+const uint32_t BUFF_STABILITY = 1122;//Stability
+const uint32_t BUFF_VAPOR_FORM = 5620;//Vapor Form
+const uint32_t BUFF_ILLUSION_OF_LIFE = 10346;//Illusion of Life
+const uint32_t BUFF_CRYSTALLINE_HEART = 37657;//Crystalline Heart, for envoy armor 2 collection
